Moved Direction class to org.usfirst.frc2183.util.unit
