+++ Robot2018/src/org/usfirst/frc2813/Robot2018/OI.java
@@ -12,6 +12,7 @@ import org.usfirst.frc2813.Robot2018.commands.Arm.SetJaws;
 import org.usfirst.frc2813.Robot2018.commands.DriveTrain.OIDrive;
 import org.usfirst.frc2813.Robot2018.commands.DriveTrain.ResetEncoders;
 import org.usfirst.frc2813.Robot2018.subsystems.RoboRIOUserButton;
+import org.usfirst.frc2813.util.unit.Direction;
 
 import edu.wpi.first.wpilibj.Compressor;
 import edu.wpi.first.wpilibj.Joystick;
