+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Arm/ArmHoldPosition.java
@@ -1,6 +1,7 @@
 package org.usfirst.frc2813.Robot2018.commands.Arm;
 
 import org.usfirst.frc2813.Robot2018.Robot;
+import org.usfirst.frc2813.Robot2018.commands.GearheadsCommand;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
 
 /**
@@ -10,7 +11,6 @@ import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
  */
 public class ArmHoldPosition extends GearheadsInstantCommand {
 	public ArmHoldPosition() {
-		logger.info("Active hold position");
 		requires(Robot.arm);
 	}
