+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Arm/ArmMoveInDirection.java
@@ -1,8 +1,8 @@
 package org.usfirst.frc2813.Robot2018.commands.Arm;
 
-import org.usfirst.frc2813.Robot2018.Direction;
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsCommand;
+import org.usfirst.frc2813.util.unit.Direction;
 
 /**
  * Move arm in given direction until interrupted.
@@ -13,7 +13,6 @@ public class ArmMoveInDirection extends GearheadsCommand {
 
 	public ArmMoveInDirection(Direction direction) {
 		this.direction = direction;
-		logger.info("Move in " + direction);
 		requires(Robot.arm);
 	}
