+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Arm/ArmMoveToPosition.java
@@ -2,6 +2,7 @@ package org.usfirst.frc2813.Robot2018.commands.Arm;
 
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
+import org.usfirst.frc2813.util.unit.values.Length;
 
 /**
  * Move arm to position. Motor controller does this without
@@ -11,7 +12,6 @@ public class ArmMoveToPosition extends GearheadsInstantCommand {
 	private final double positionInInches;
 
 	public ArmMoveToPosition(double positionInInches) {
-		logger.info("Move to position: " + positionInInches);
 		this.positionInInches = positionInInches;
 		requires(Robot.arm);
 	}
