+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Arm/SetJaws.java
@@ -4,8 +4,8 @@ package org.usfirst.frc2813.Robot2018.commands.Arm;
 
 import edu.wpi.first.wpilibj.command.InstantCommand;
 
-import org.usfirst.frc2813.Robot2018.Direction;
 import org.usfirst.frc2813.Robot2018.Robot;
+import org.usfirst.frc2813.util.unit.Direction;
 
 /**
  * Open/Close the Robot arm jaws
