+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Arm/SpinIntake.java
@@ -4,9 +4,9 @@
 package org.usfirst.frc2813.Robot2018.commands.Arm;
 import edu.wpi.first.wpilibj.command.Command;
 
-import org.usfirst.frc2813.Robot2018.Direction;
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.subsystems.Arm;
+import org.usfirst.frc2813.util.unit.Direction;
 
 /**
  * Spin or halt the Robot arm intake
