+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Elevator/ElevatorHoldPosition.java
@@ -1,6 +1,7 @@
 package org.usfirst.frc2813.Robot2018.commands.Elevator;
 
 import org.usfirst.frc2813.Robot2018.Robot;
+import org.usfirst.frc2813.Robot2018.commands.GearheadsCommand;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
 
 /**
@@ -10,7 +11,6 @@ import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
  */
 public class ElevatorHoldPosition extends GearheadsInstantCommand {
 	public ElevatorHoldPosition() {
-		logger.info("Active hold position");
 		requires(Robot.elevator);
 	}
