+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Elevator/ElevatorMoveInDirection.java
@@ -1,8 +1,8 @@
 package org.usfirst.frc2813.Robot2018.commands.Elevator;
 
-import org.usfirst.frc2813.Robot2018.Direction;
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsCommand;
+import org.usfirst.frc2813.util.unit.Direction;
 
 /**
  * Move elevator in given direction until interrupted.
@@ -20,7 +20,7 @@ public class ElevatorMoveInDirection extends GearheadsCommand {
 	// @Override
 	protected void initialize() {
 		logger.finer("in initialize");
-			Robot.elevator.moveInDirection(direction);
+		Robot.elevator.moveInDirection(direction);
 	}
 
 	//@Override
