+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Elevator/ElevatorMoveInDirectionAtSpeed.java
@@ -1,8 +1,9 @@
 package org.usfirst.frc2813.Robot2018.commands.Elevator;
 
-import org.usfirst.frc2813.Robot2018.Direction;
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsCommand;
+import org.usfirst.frc2813.util.unit.Direction;
+import org.usfirst.frc2813.util.unit.values.Rate;
 
 /**
  * Move elevator in given direction at given speed until interrupted.
