+++ Robot2018/src/org/usfirst/frc2813/Robot2018/subsystems/SubsystemDirectionSpeed.java
@@ -1,6 +1,6 @@
 package org.usfirst.frc2813.Robot2018.subsystems;
 
-import org.usfirst.frc2813.Robot2018.Direction;
+import org.usfirst.frc2813.util.unit.Direction;
 
 
 /**
