+++ Robot2018/src/org/usfirst/frc2813/Robot2018/subsystems/SubsystemPositionDirectionSpeed.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc2813.Robot2018.subsystems;
 
-import org.usfirst.frc2813.Robot2018.Direction;
 import org.usfirst.frc2813.Robot2018.MotorControllerState;
+import org.usfirst.frc2813.util.unit.Direction;
 
 /**
  * Subsystem baseclass for subsystems which move in two direction
